pyproject.toml
[tool.poetry]
name = "shortener-fastapi"
version = "0.1.0"
description = "Simple URL shortener service. Python 3 + FastAPI"
authors = ["Kochurov Mikhail <firepanda70@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic-settings = "^2.1.0"
asyncpg = "^0.29.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
mixer = "^7.2.2"
aiosqlite = "^0.19.0"
pytest-freezer = "^0.4.8"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
